@@ -16,7 +16,7 @@
       "issues": "https://github.com/flextype-plugins/icon/issues"
     },
     "require": {
-        "php": ">=7.3.0",
+        "php": ">=7.4.0",
         "flextype-components/arrays" : "3.0.1",
         "flextype-components/filesystem": "2.0.8"
     },
@@ -24,7 +24,7 @@
         "apcu-autoloader": true,
         "optimize-autoloader": true,
         "platform": {
-            "php": "7.3.0"
+            "php": "7.4.0"
         }
     },
     "autoload": {
