@@ -22,52 +22,41 @@ use Twig\Extension\DebugExtension;
 /**
  * Add CSRF (cross-site request forgery) protection service to Flextype container
  */
-flextype()->container()['csrf'] = static function (){
-    return new Guard();
-};
+flextype()->container()['csrf'] = fn() => new Guard();
 
 /**
  * Add Twig service to Flextype container
  */
-flextype()->container()['twig'] = static function () {
-
-    // Get twig settings
-    $twigSettings = [
-                     'auto_reload' => flextype('registry')->get('plugins.twig.settings.auto_reload'),
-                     'cache' => flextype('registry')->get('plugins.twig.settings.cache') ? PATH['tmp'] . '/twig' : false,
-                     'debug' => flextype('registry')->get('plugins.twig.settings.debug'),
-                     'charset' => flextype('registry')->get('plugins.twig.settings.charset')
-                    ];
+flextype()->container()['twig'] = function () {
 
     // Create Twig View
-    $twig = new Twig(PATH['project'], $twigSettings);
-
-    // Instantiate
-    $router = flextype('router');
-    $uri    = Uri::createFromEnvironment(new Environment($_SERVER));
+    $twig = new Twig(PATH['project'],
+                        ['auto_reload' => flextype('registry')->get('plugins.twig.settings.auto_reload'),
+                         'cache' => flextype('registry')->get('plugins.twig.settings.cache') ? PATH['tmp'] . '/twig' : false,
+                         'debug' => flextype('registry')->get('plugins.twig.settings.debug'),
+                         'charset' => flextype('registry')->get('plugins.twig.settings.charset')]);
 
     // Add Twig Extension
-    $twig->addExtension(new TwigExtension($router, $uri));
+    $twig->addExtension(new TwigExtension(flextype('router'),
+                                          Uri::createFromEnvironment(new Environment($_SERVER))));
 
     // Add Twig Debug Extension
     $twig->addExtension(new DebugExtension());
 
     // Load Flextype Twig extensions from directory /flextype/twig/ based on settings.twig.extensions array
-    $twig_extensions = flextype('registry')->get('plugins.twig.settings.extensions');
+    $twigExtensions = flextype('registry')->get('plugins.twig.settings.extensions');
 
-    foreach ($twig_extensions as $twig_extension) {
-        $twig_extension_class_name = $twig_extension . 'TwigExtension';
-        $twig_extension_class_name_with_namespace = 'Flextype\\Plugin\\Twig\\Twig\\' . $twig_extension . 'TwigExtension';
+    foreach ($twigExtensions as $twigExtension) {
+        $twigExtensionClassName = $twigExtension . 'TwigExtension';
+        $twigExtensionClassNameWithNamespace = 'Flextype\\Plugin\\Twig\\Twig\\' . $twigExtension . 'TwigExtension';
 
-        if (file_exists(ROOT_DIR . '/project/plugins/twig/twig/' . $twig_extension_class_name . '.php')) {
+        if (file_exists(ROOT_DIR . '/project/plugins/twig/twig/' . $twigExtensionClassName . '.php')) {
 
-            if ($twig_extension == 'Flash') {
-                flextype()->container()['flash'] = static function () {
-                    return new Messages();
-                };
+            if ($twigExtension == 'Flextype') {
+                flextype()->container()['flash'] = fn() => new Messages();
             }
 
-            $twig->addExtension(new $twig_extension_class_name_with_namespace());
+            $twig->addExtension(new $twigExtensionClassNameWithNamespace());
         }
     }
 
