@@ -7,7 +7,7 @@ namespace Flextype\Plugin\Sitemap;
  * Flextype Sitemap Plugin
  *
  * @author Romanenko Sergey / Awilum <awilum@yandex.ru>
- * @link http://flextype.org
+ * @link https://flextype.org
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
